requirements.txt
python-dotenv
streamlit
sentence-transformers
langchain-community
langchain-classic
langchain
langchain-huggingface
langchain-text_splitters
unstructured
unstructured[pdf]
langchain-unstructured
langchain-chroma
langchain-groq